+++ frontend/src/index.js
@@ -1,12 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import FrontPage from './FrontPage';
 import App from './App';
 
 ReactDOM.render(
   <React.StrictMode>
-    <FrontPage />
+    <App />
   </React.StrictMode>,
   document.getElementById('root')
 );
